modules/translator_dev/requirements.txt
transformers>=4.36.0
torch>=2.1.0
sentencepiece>=0.1.99
diskcache>=5.6.3
pyyaml>=6.0.1
tqdm>=4.66.1
colorama>=0.4.6
rich>=13.7.0
openai>=1.10.0
langdetect>=1.0.9
argparse>=1.4.0
beautifulsoup4>=4.12.2 